Kdf group (#16)
* 1.15.7+4184
fix: prekey message to send cashu or image

* fix: receive all pendings cashu bills

* feat: add kdf group

* kdf group init

* kdf group hello message

* feat hello message

* send kdf message

* admin create room and join

* feat: create account by same mnemonic

* feat: support multi signalId

* update room info

* fix

* prekey message

* Connect bug fix (#21)

* #1.16.2+4215
1. fix: websocket reconnet
2. fix: wallet not init error

* fix

* Connect bug fix (#21)

* #1.16.2+4215
1. fix: websocket reconnet
2. fix: wallet not init error

* fix

* feat: store prikey to keychain

* 1.16.3+4217
1. feat: store prikey to keychain. Migrate old version prikey data
2. fix: auto delete chat room controller. Solve memory leak problem

* fix: init ecash when create the first account

* fix

* fix: init ecash when create the first account

* fix

* fix

* Keychain (#25)

* feat: store prikey to keychain

* 1.16.3+4217
1. feat: store prikey to keychain. Migrate old version prikey data
2. fix: auto delete chat room controller. Solve memory leak problem

* fix: init ecash when create the first account

* fix

* 1.16.3+4219

* Keychain merge to dev branch (#24)

* feat: store prikey to keychain

* 1.16.3+4217
1. feat: store prikey to keychain. Migrate old version prikey data
2. fix: auto delete chat room controller. Solve memory leak problem

* fix: init ecash when create the first account

* fix

---------


---------


* 1.16.3+4217  (#23)

* feat: store prikey to keychain

* 1.16.3+4217
1. feat: store prikey to keychain. Migrate old version prikey data
2. fix: auto delete chat room controller. Solve memory leak problem

* fix: init ecash when create the first account

* fix

* Keychain (#27)

* feat: store prikey to keychain

* 1.16.3+4217
1. feat: store prikey to keychain. Migrate old version prikey data
2. fix: auto delete chat room controller. Solve memory leak problem

* fix: init ecash when create the first account

* fix

* 1.16.3+4219

* Keychain merge to dev branch (#24)

* feat: store prikey to keychain

* 1.16.3+4217
1. feat: store prikey to keychain. Migrate old version prikey data
2. fix: auto delete chat room controller. Solve memory leak problem

* fix: init ecash when create the first account

* fix

---------


* 1.16.3+4217  (#23)

* feat: store prikey to keychain

* 1.16.3+4217
1. feat: store prikey to keychain. Migrate old version prikey data
2. fix: auto delete chat room controller. Solve memory leak problem

* fix: init ecash when create the first account

* fix

---------


* add not prekey message

* feat: send message wrap with nip4 event, it used to protect sender's pubkey

* feat: kdf group

* feat: support medium group, encrypt message with signal protocol

* text

* fix

* style

* feat: send cashu with sendStamp
fix: style of message fontSize
config: set  prekey message count to 2

* 1.17.2+4223
style: fontsize and color for message widget

* feat: admin add member to group

* fix: warning

* feat: request to invite user to join group

* feat: remove member from group

* feat: remove member and reinvite exist members

* feat: use a new signalId to join each kdf group

* 1.18.1+4224
feat: 1. clear keys when expired
feat: 2. Check if sender is an administrator for key operations
feat: 3. more setting pages support clear unread messages

* reset session debug (#30)

* add send cashu with sendStamp

* Merge

* fix

* reset session debug (#30)

* add send cashu with sendStamp

* fix

* remove a

---------

+++ packages/app/lib/service/secure_storage.dart
@@ -1,4 +1,6 @@
+import 'package:flutter/foundation.dart';
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
+import 'package:app/utils/config.dart';
 
 class SecureStorage {
   static SecureStorage? _instance;
@@ -8,7 +10,7 @@ class SecureStorage {
   static FlutterSecureStorage storage = const FlutterSecureStorage(
       iOptions: IOSOptions(accessibility: KeychainAccessibility.first_unlock));
 
-  String mnemonicKey = 'mnemonic';
+  String mnemonicKey = kReleaseMode ? 'mnemonic' : '${Config.env}:mnemonic';
   static Map<String, String> keys = {};
 
   Future writePhraseWords(String words) async {
@@ -47,7 +49,12 @@ class SecureStorage {
     throw Exception('$pubkey \'s private key not found');
   }
 
-  String _getPrivateKeyName(String pubkey) => "prikey:$pubkey";
+  String _getPrivateKeyName(String pubkey) {
+    if (kReleaseMode) {
+      return "prikey:$pubkey";
+    }
+    return "${Config.env}:prikey:$pubkey";
+  }
 
   Future<Map<String, String>> readAll() {
     return storage.readAll();
@@ -58,9 +65,12 @@ class SecureStorage {
     await storage.delete(key: key);
   }
 
-  Future clearAll() async {
-    await storage.deleteAll(
-        iOptions: const IOSOptions(
-            accessibility: KeychainAccessibility.first_unlock));
+  // in debug model, if you open multi clients, all of them share the same keychain
+  Future clearAll([bool force = false]) async {
+    if (force || kReleaseMode) {
+      await storage.deleteAll(
+          iOptions: const IOSOptions(
+              accessibility: KeychainAccessibility.first_unlock));
+    }
   }
 }
